--- java/simple-with-process-resources-spec/app.json
@@ -1,8 +1,0 @@
-{
-  "name": "simple-with-process",
-  "language": "java",
-  "environment": "common",
-  "resources": {
-    "source_name": "fixtures/demo-cdc.json"
-  }
-}
