--- java/simple-with-process-resources-spec/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-quarkus.package.main-class=com.meroxa.turbine.Main
