+++ java/simple-with-process/fixtures/demo-cdc.json
