+++ java/simple-with-process/fixtures/demo-no-cdc.json
