+++ java/simple-with-process/src/main/resources/application.properties
